+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 * Add `Date#timestamp` getter
 * Add `flag` parameter to `Object.path(flag, obj, path)`
 * Add `Blast.PATH_AGGREGATE` flag
+* Add flag support to `Array#sortByPath()`
 
 ## 0.6.2 (2018-11-09)